refactoring(StepsPayments):create component and refactoring DetailPresentation
